backend/management/requirements.txt
asgiref==3.5.2
cffi==1.15.1
Django==4.1.4
djangorestframework==3.14.0
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
pycparser==2.21
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
zope.event==4.6
zope.interface==5.5.2
